extracted procedure component
